python/requirements.txt
# SAPSE-Synergy Python Dependencies
# Python >= 3.10 required

# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
openai>=1.0.0

# Vector search
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
loguru>=0.7.0
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
